requirements.txt
requests==2.32.3
minio==7.2.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1